routing: move payment session to separate file
